requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
PyMuPDF>=1.22.0
opencv-python>=4.7.0
pytesseract>=0.3.10
Pillow>=9.5.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
albumentations>=1.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
